Update test_physics.cpp
I changed the inputs to particles as we agreed (Particles(filename, nparticles)), keeping the outputs the way  simulation and physics would like them to be. In order to commit the changes, I have to update the Particles calls in the test_physics file. (to Particles *particles = new Particles(filename, nparticles)).